+++ packages/yarnpkg-pnp/sources/esm-loader/loaderUtils.ts
@@ -16,9 +16,9 @@ export async function tryReadFile(path: NativePath): Promise<string | null> {
   }
 }
 
-export function tryParseURL(str: string) {
+export function tryParseURL(str: string, base?: string | URL | undefined) {
   try {
-    return new URL(str);
+    return new URL(str, base);
   } catch {
     return null;
   }